Clean up WKB test cases
